create LatticeHilbert
