+++ netket/hilbert/abstract_hilbert.py
@@ -14,7 +14,7 @@
 
 import abc
 
-from typing import List, Tuple, Optional, Union, Iterable, Iterator
+from typing import Tuple, Optional, Union, Iterable
 
 import jax.numpy as jnp
 import numpy as np
@@ -69,20 +69,17 @@ class AbstractHilbert(abc.ABC):
         raise NotImplementedError()  # pragma: no cover
 
     @property
-    @abc.abstractmethod
     def is_discrete(self) -> bool:
         r"""Whether the hilbert space is discrete."""
-        raise NotImplementedError()  # pragma: no cover
+        return False  # pragma: no cover
 
     @property
-    @abc.abstractmethod
     def is_finite(self) -> bool:
         r"""Whether the local hilbert space is finite."""
-        raise NotImplementedError()  # pragma: no cover
+        return False  # pragma: no cover
 
-    @abc.abstractmethod
     def size_at_index(self, i: int) -> int:
-        r"""Size of the local degrees of freedom at the site i.
+        r"""Size of the local degrees of freedom for the i-th variable.
 
         Args:
             i: The index of the desired site
@@ -90,93 +87,13 @@ class AbstractHilbert(abc.ABC):
         Returns:
             The number of degrees of freedom at that site
         """
-        raise NotImplementedError()  # pragma: no cover
-
-    @abc.abstractmethod
-    def states_at_index(self, i: int) -> Optional[List[float]]:
-        r"""A list of discrete local quantum numbers at the site i.
-        If the local states are infinitely many, None is returned.
-
-        Args:
-            i: The index of the desired site.
-
-        Returns:
-            A list of values or None if there are infintely many.
-        """
-        raise NotImplementedError()  # pragma: no cover
-
-    def numbers_to_states(
-        self, numbers: Union[int, np.ndarray], out: Optional[np.ndarray] = None
-    ) -> np.ndarray:
-        r"""Returns the quantum numbers corresponding to the n-th basis state
-        for input n. n is an array of integer indices such that
-        :code:py:`numbers[k]=Index(states[k])`.
-        Throws an exception iff the space is not indexable.
-
-        Args:
-            numbers (numpy.array): Batch of input numbers to be converted into arrays of
-                quantum numbers.
-            out: Optional Array of quantum numbers corresponding to numbers.
-        """
-        if out is None:
-            out = np.empty((np.atleast_1d(numbers).shape[0], self.size))
-
-        if np.any(numbers >= self.n_states):
-            raise ValueError("numbers outside the range of allowed states")
-
-        if np.isscalar(numbers):
-            return self._numbers_to_states(np.atleast_1d(numbers), out=out)[0, :]
-        else:
-            return self._numbers_to_states(numbers, out=out)
-
-    def states_to_numbers(
-        self, states: np.ndarray, out: Optional[np.ndarray] = None
-    ) -> Union[int, np.ndarray]:
-        r"""Returns the basis state number corresponding to given quantum states.
-        The states are given in a batch, such that states[k] has shape (hilbert.size).
-        Throws an exception iff the space is not indexable.
-
-        Args:
-            states: Batch of states to be converted into the corresponding integers.
-            out: Array of integers such that out[k]=Index(states[k]).
-                 If None, memory is allocated.
-
-        Returns:
-            numpy.darray: Array of integers corresponding to out.
-        """
-        if states.shape[-1] != self.size:
-            raise ValueError(
-                f"Size of this state ({states.shape[-1]}) not"
-                f"corresponding to this hilbert space {self.size}"
-            )
-
-        states_r = np.asarray(np.reshape(states, (-1, states.shape[-1])))
-
-        if out is None:
-            out = np.empty(states_r.shape[:-1], dtype=np.int64)
-
-        out = self._states_to_numbers(states_r, out=out.reshape(-1))
-
-        if states.ndim == 1:
-            return out[0]
-        else:
-            return out.reshape(states.shape[:-1])
+        return self.shape[i]  # prgma: no cover
 
     @property
     def n_states(self) -> int:
         r"""The total dimension of the many-body Hilbert space.
         Throws an exception iff the space is not indexable."""
-        raise NotImplementedError()
-
-    def states(self) -> Iterator[np.ndarray]:
-        r"""Returns an iterator over all valid configurations of the Hilbert space.
-        Throws an exception iff the space is not indexable.
-        Iterating over all states with this method is typically inefficient,
-        and ```all_states``` should be prefered.
-
-        """
-        for i in range(self.n_states):
-            yield self.numbers_to_states(i).reshape(-1)
+        raise ValueError("The Hilbert space is infinite.")
 
     # after removing legacy:
     # signature must be the following
@@ -239,22 +156,6 @@ class AbstractHilbert(abc.ABC):
 
             return random.random_state(self, key, size, dtype=dtype)
 
-    def all_states(self, out: Optional[np.ndarray] = None) -> np.ndarray:
-        r"""Returns all valid states of the Hilbert space.
-
-        Throws an exception if the space is not indexable.
-
-        Args:
-            out: an optional pre-allocated output array
-
-        Returns:
-            A (n_states x size) batch of statess. this corresponds
-            to the pre-allocated array if it was passed.
-        """
-        numbers = np.arange(0, self.n_states, dtype=np.int64)
-
-        return self.numbers_to_states(numbers, out)
-
     def ptrace(self, sites: Union[int, Iterable]) -> "AbstractHilbert":
         """Returns the hilbert space without the selected sites.
 
@@ -267,33 +168,12 @@ class AbstractHilbert(abc.ABC):
             The partially-traced hilbert space. The type of the resulting hilbert space
             might be different from the starting one.
         """
-        raise NotImplementedError("Ptrace not implemented for this hilbert space type")
+        raise NotImplementedError("Ptrace not implemented for this hilbert space type.")
 
     @property
     def is_indexable(self) -> bool:
         """Whever the space can be indexed with an integer"""
-        if not self.is_discrete:
-            return False
-
-        if not self.is_finite:
-            return False
-
-        log_max = np.log(max_states)
-
-        return np.sum(np.log(self.shape)) <= log_max
-
-    def __mul__(self, other: "AbstractHilbert"):
-        if self == other:
-            return self ** 2
-        else:
-            from .tensor_hilbert import TensorHilbert
-
-            if type(self) == type(other):
-                res = self._mul_sametype_(other)
-                if res is not NotImplemented:
-                    return res
-
-            return TensorHilbert(self) * other
+        return False
 
     @property
     @abc.abstractmethod
